通常のautoware_launchを立ち上げた後に、データ収集用launchをもう一つ立ち上げるように変更
