+++ control_data_collecting_tool/README.md
@@ -4,25 +4,24 @@ This package provides a tool for collecting data by pure pursuit control within 
 
 ## 使い方（開発中バージョン）
 
-0 public版のautowareと、このパッケージを含むリポジトリをgit clone
-
-1 colcon build
-
-2 autoware立ち上げ
+1 launch autoware
 
 ```bash
-ros2 launch control_data_collecting_tool planning_simulator.launch.xml map_path:=$HOME/autoware_map/sample-map-planning vehicle_model:=sample_vehicle sensor_model:=sample_sensor_kit trajectory_follower_mode:=control_data_collecting_tool
+ros2 launch autoware_launch planning_simulator.launch.xml rviz_config:=$(ros2 pkg prefix control_data_collecting_tool)/share/control_data_collecting_tool/rviz/autoware.rviz map_path:=$HOME/autoware_map/sample-map-planning vehicle_model:=sample_vehicle sensor_model:=sample_sensor_kit
 ```
 
+2 launch control_data_collecting_tool
+```bash
+ros2 launch control_data_collecting_tool control_data_collecting_tool.launch.py
+```
 3 `DataCollectingAreaSelectionTool`を使ってデータ収集領域を選択
 
 - 走行中にはデータ収集領域を変更できない
 
 4 `2D Pose Estimate`を使って initial poseを置く
 
-5 `LOCAL`を押して走行を開始する（現在は目標軌道は適当に仮置きしている）
+5 `LOCAL`を押して走行を開始する
 
 ## TODO
-
 - ドキュメント整備
 - 不要な変数、パッケージ、ヘッダーの削除
