--- control_data_collecting_tool/launch/planning_simulator.launch.xml
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<launch>
-  <!-- Essential parameters -->
-  <arg name="map_path" description="point cloud and lanelet2 map directory path"/>
-  <arg name="vehicle_model" default="sample_vehicle" description="vehicle model name"/>
-  <arg name="sensor_model" default="sample_sensor_kit" description="sensor model name"/>
-  <arg name="data_path" default="$(env HOME)/autoware_data" description="packages data and artifacts directory path"/>
-
-  <!-- launch module preset -->
-  <arg name="planning_module_preset" default="default" description="planning module preset"/>
-
-  <!-- Optional parameters -->
-  <!-- Map -->
-  <arg name="lanelet2_map_file" default="lanelet2_map.osm" description="lanelet2 map file name"/>
-  <arg name="pointcloud_map_file" default="pointcloud_map.pcd" description="pointcloud map file name"/>
-  <let name="rviz_initial_pose_auto_fix_target" value="vector_map"/>
-  <let name="gnss_initial_pose_auto_fix_target" value="vector_map"/>
-  <!-- System -->
-  <arg name="launch_dummy_doors" default="true" description="launch dummy doors simulation"/>
-  <arg name="launch_system_monitor" default="false" description="launch system monitor"/>
-  <arg name="launch_dummy_diag_publisher" default="false" description="launch dummy diag publisher"/>
-  <!-- Scenario simulation -->
-  <arg name="initial_engage_state" default="true" description="/vehicle/engage state after starting Autoware"/>
-  <arg name="perception/enable_detection_failure" default="true" description="enable to simulate detection failure when using dummy perception"/>
-  <arg name="perception/enable_object_recognition" default="true" description="enable object recognition when using dummy perception"/>
-  <arg name="perception/use_base_link_z" default="true" description="dummy perception uses base_link z axis coordinate if it is true"/>
-  <arg name="sensing/visible_range" default="300.0" description="visible range when using dummy perception"/>
-  <arg name="scenario_simulation" default="false" description="use scenario simulation"/>
-  <!-- Tools -->
-  <arg name="rviz" default="true" description="launch rviz"/>
-  <arg name="rviz_config" default="$(find-pkg-share control_data_collecting_tool)/rviz/autoware.rviz" description="rviz config"/>
-  <let name="rviz_respawn" value="false" if="$(var scenario_simulation)"/>
-  <let name="rviz_respawn" value="true" unless="$(var scenario_simulation)"/>
-  <!-- Vcu emulation -->
-  <arg name="vehicle_simulation" default="true" description="use vehicle simulation"/>
-  <!-- Auto mode setting-->
-  <arg name="enable_all_modules_auto_mode" default="$(var scenario_simulation)" description="enable all module's auto mode"/>
-  <!-- Simulated time -->
-  <arg name="use_sim_time" default="$(var scenario_simulation)"/>
-
-  <group scoped="false">
-    <!-- Vehicle -->
-    <let name="launch_vehicle_interface" value="false" if="$(var vehicle_simulation)"/>
-    <let name="launch_vehicle_interface" value="true" unless="$(var vehicle_simulation)"/>
-
-    <include file="$(find-pkg-share control_data_collecting_tool)/launch/autoware.launch.xml">
-      <!-- Common -->
-      <arg name="map_path" value="$(var map_path)"/>
-      <arg name="vehicle_model" value="$(var vehicle_model)"/>
-      <arg name="sensor_model" value="$(var sensor_model)"/>
-      <arg name="data_path" value="$(var data_path)"/>
-      <!-- launch module preset -->
-      <arg name="planning_module_preset" value="$(var planning_module_preset)"/>
-      <!-- Modules to be launched -->
-      <arg name="launch_sensing" value="false"/>
-      <arg name="launch_localization" value="false"/>
-      <arg name="launch_perception" value="false"/>
-      <!-- Vehicle -->
-      <arg name="launch_vehicle_interface" value="$(var launch_vehicle_interface)"/>
-      <!-- System -->
-      <arg name="system_run_mode" value="planning_simulation"/>
-      <arg name="launch_system_monitor" value="$(var launch_system_monitor)"/>
-      <arg name="launch_dummy_diag_publisher" value="$(var launch_dummy_diag_publisher)"/>
-      <!-- Map -->
-      <arg name="lanelet2_map_file" value="$(var lanelet2_map_file)"/>
-      <arg name="pointcloud_map_file" value="$(var pointcloud_map_file)"/>
-      <!-- Tools -->
-      <arg name="rviz" value="$(var rviz)"/>
-      <arg name="rviz_config" value="$(var rviz_config)"/>
-      <arg name="rviz_respawn" value="$(var rviz_respawn)"/>
-      <!-- Auto mode setting-->
-      <arg name="enable_all_modules_auto_mode" value="$(var enable_all_modules_auto_mode)"/>
-      <arg name="is_simulation" value="true"/>
-    </include>
-  </group>
-
-  <!-- Simulator -->
-  <group>
-    <let name="launch_dummy_perception" value="false" if="$(var scenario_simulation)"/>
-    <let name="launch_dummy_perception" value="true" unless="$(var scenario_simulation)"/>
-    <let name="launch_dummy_vehicle" value="false" if="$(var scenario_simulation)"/>
-    <let name="launch_dummy_vehicle" value="true" unless="$(var scenario_simulation)"/>
-    <let name="launch_dummy_localization" value="true"/>
-    <let name="launch_diagnostic_converter" value="$(var scenario_simulation)"/>
-
-    <include file="$(find-pkg-share autoware_launch)/launch/components/tier4_simulator_component.launch.xml">
-      <arg name="launch_dummy_perception" value="$(var launch_dummy_perception)"/>
-      <arg name="launch_dummy_vehicle" value="$(var launch_dummy_vehicle)"/>
-      <arg name="launch_dummy_localization" value="$(var launch_dummy_localization)"/>
-      <arg name="launch_dummy_doors" value="$(var launch_dummy_doors)"/>
-      <arg name="launch_diagnostic_converter" value="$(var launch_diagnostic_converter)"/>
-      <arg name="perception/enable_detection_failure" value="$(var perception/enable_detection_failure)"/>
-      <arg name="perception/enable_object_recognition" value="$(var perception/enable_object_recognition)"/>
-      <arg name="perception/use_base_link_z" value="$(var perception/use_base_link_z)"/>
-      <arg name="sensing/visible_range" value="$(var sensing/visible_range)"/>
-      <arg name="vehicle_model" value="$(var vehicle_model)"/>
-      <arg name="initial_engage_state" value="$(var initial_engage_state)"/>
-      <arg name="vehicle_info_param_file" value="$(find-pkg-share $(var vehicle_model)_description)/config/vehicle_info.param.yaml"/>
-    </include>
-  </group>
-</launch>
