--- control_data_collecting_tool/launch/tier4_control_component.launch.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-  <!-- optional parameters are written here -->
-  <arg name="lateral_controller_mode" default="mpc"/>
-  <arg name="longitudinal_controller_mode" default="pid"/>
-  <arg name="use_individual_control_param" default="false"/>
-  <arg name="enable_autonomous_emergency_braking" default="true"/>
-  <arg name="enable_predicted_path_checker" default="false"/>
-
-  <let name="latlon_controller_param_path_dir" value="$(var vehicle_id)" if="$(var use_individual_control_param)"/>
-  <let name="latlon_controller_param_path_dir" value="" unless="$(var use_individual_control_param)"/>
-
-  <include file="$(find-pkg-share control_data_collecting_tool)/launch/control.launch.py">
-    <!-- option -->
-    <arg name="vehicle_param_file" value="$(find-pkg-share $(var vehicle_model)_description)/config/vehicle_info.param.yaml"/>
-    <arg name="vehicle_id" value="$(var vehicle_id)"/>
-    <arg name="enable_obstacle_collision_checker" value="false"/>
-    <arg name="trajectory_follower_node_param_path" value="$(find-pkg-share autoware_launch)/config/control/trajectory_follower/trajectory_follower_node.param.yaml"/>
-    <arg
-      name="lat_controller_param_path"
-      value="$(find-pkg-share autoware_launch)/config/control/trajectory_follower/$(var latlon_controller_param_path_dir)/lateral/$(var lateral_controller_mode).param.yaml"
-    />
-    <arg
-      name="lon_controller_param_path"
-      value="$(find-pkg-share autoware_launch)/config/control/trajectory_follower/$(var latlon_controller_param_path_dir)/longitudinal/$(var longitudinal_controller_mode).param.yaml"
-    />
-    <arg name="lateral_controller_mode" value="$(var lateral_controller_mode)"/>
-    <arg name="longitudinal_controller_mode" value="$(var longitudinal_controller_mode)"/>
-    <arg name="check_external_emergency_heartbeat" value="$(var check_external_emergency_heartbeat)"/>
-
-    <!-- common param path -->
-    <arg name="nearest_search_param_path" value="$(find-pkg-share autoware_launch)/config/control/common/nearest_search.param.yaml"/>
-
-    <!-- package param path -->
-    <arg name="vehicle_cmd_gate_param_path" value="$(find-pkg-share autoware_launch)/config/control/vehicle_cmd_gate/vehicle_cmd_gate.param.yaml"/>
-    <arg name="lane_departure_checker_param_path" value="$(find-pkg-share autoware_launch)/config/control/lane_departure_checker/lane_departure_checker.param.yaml"/>
-    <arg name="control_validator_param_path" value="$(find-pkg-share autoware_launch)/config/control/control_validator/control_validator.param.yaml"/>
-    <arg name="operation_mode_transition_manager_param_path" value="$(find-pkg-share autoware_launch)/config/control/operation_mode_transition_manager/operation_mode_transition_manager.param.yaml"/>
-    <arg name="shift_decider_param_path" value="$(find-pkg-share autoware_launch)/config/control/shift_decider/shift_decider.param.yaml"/>
-    <arg name="obstacle_collision_checker_param_path" value="$(find-pkg-share autoware_launch)/config/control/obstacle_collision_checker/obstacle_collision_checker.param.yaml"/>
-    <arg name="external_cmd_selector_param_path" value="$(find-pkg-share autoware_launch)/config/control/external_cmd_selector/external_cmd_selector.param.yaml"/>
-    <arg name="aeb_param_path" value="$(find-pkg-share autoware_launch)/config/control/autonomous_emergency_braking/autonomous_emergency_braking.param.yaml"/>
-    <arg name="enable_autonomous_emergency_braking" value="$(var enable_autonomous_emergency_braking)"/>
-    <arg name="enable_predicted_path_checker" value="$(var enable_predicted_path_checker)"/>
-    <arg name="predicted_path_checker_param_path" value="$(find-pkg-share autoware_launch)/config/control/predicted_path_checker/predicted_path_checker.param.yaml"/>
-  </include>
-</launch>
